+++ app/controllers/resources_controller.rb
@@ -45,16 +45,30 @@ class ResourcesController < RestController
     end
 
     result =
-      if params[:count] == 'true'
-        { count: scope.count('*'.lit) }
+    if params[:count] == 'true'
+      { count: scope.count('*'.lit) }
+    else
+      role = assumed_role(query_role)
+      # all_roles- find all role memberships, expanded recursively.
+      role_membership = role.all_roles.map(&:role_id)
+      # Get all resources
+      result = scope.select(:resources.*)
+      # If we want only the short result map only the resource ids
+      if params[:short] == 'true'
+        # Map only the resources id
+        result = result.select(:resource_id).all.map { |obj| obj.id }
       else
-        scope.select(:resources.*).
+        # Add information about the resources
+        result = result.
           eager(:annotations).
-          eager(:permissions).
+          eager(:permissions => lambda { |permission| permission.where(role_id: role_membership) }).
           eager(:secrets).
-          eager(:policy_versions).
-          all
+          eager(:policy_versions)
+        # run the query
+        result = result.all
       end
+      result
+    end
     audit_list_success(options)
     render(json: result)
   end