+++ app/domain/edge_logic/data_handlers/ongoing_handler.rb
@@ -11,7 +11,7 @@ module EdgeLogic
 
       def call(params, hostname, ip)
         params.require(:edge_statistics).require(:last_synch_time)
-        allowed_params = [:edge_version, :edge_container_type, edge_statistics: [:last_synch_time, cycle_requests:
+        allowed_params = [:edge_version, :edge_container_type, :edge_proxy, edge_statistics: [:last_synch_time, cycle_requests:
           [:get_secret, :apikey_authenticate, :jwt_authenticate, :redirect]]]
         options = params.permit(*allowed_params).to_h
         validate_params(options, input_validator)
@@ -22,13 +22,19 @@ module EdgeLogic
         # Log Edge statistics to be collected by Datadog
         stats = options['edge_statistics']
         cycle_reqs = stats['cycle_requests'] || {}
+        # Log whether edge uses proxy
+        edge_proxy = options['edge_proxy']
+        edge_proxy_usage = ""
+        if edge_proxy
+          edge_proxy_usage = " using edge proxy"
+        end
         #convert time to seconds 
         last_synch_time_sec = Rational(stats['last_synch_time'], 1000)
         installation_time_sec = Rational(edge.installation_date, 1000)
         @logger.info(LogMessages::Edge::EdgeTelemetry.new(tenant_id, edge.name, Time.at(last_synch_time_sec),
                                                          cycle_reqs['get_secret'], cycle_reqs['apikey_authenticate'],
                                                          cycle_reqs['jwt_authenticate'], cycle_reqs['redirect'],
-                                                         edge.version, edge.platform, Time.at(installation_time_sec)))
+                                                         edge.version, edge.platform, Time.at(installation_time_sec), edge_proxy_usage))
       end
 
       def input_validator
@@ -38,6 +44,8 @@ module EdgeLogic
             v.match?(/^[0-9.v]+$/)
           when :edge_container_type
             %w[podman docker].include?(v.downcase)
+          when :edge_proxy
+            [true, false].include?(v)
           when :last_synch_time, :cycle_requests, :get_secret, :apikey_authenticate, :jwt_authenticate, :redirect
             numeric_validator.call(k, v)
           else