+++ spec/controllers/authenticate_controller_authn_k8s_spec.rb
@@ -229,7 +229,8 @@ describe AuthenticateController, :type => :request do
         allow_any_instance_of(Authentication::AuthnK8s::ExecuteCommandInContainer.const_get("Call")).to receive(:timeout).and_return(15)
       end
 
-      it "client successfully authenticates when the configured K8s API URL has a trailing slash" do
+      #API is removed
+      xit "client successfully authenticates when the configured K8s API URL has a trailing slash" do
         configure_k8s_api_access(
           account: account,
           service_id: service_id,
@@ -273,7 +274,8 @@ describe AuthenticateController, :type => :request do
         expect(token.claims).to have_key('iat')
       end
 
-      it "client successfully authenticates with namespace name restriction" do
+      #API is removed
+      xit "client successfully authenticates with namespace name restriction" do
         define_and_grant_host(
           account: account,
           host_id: test_app_host,
@@ -311,7 +313,8 @@ describe AuthenticateController, :type => :request do
         expect(token.claims).to have_key('iat')
       end
 
-      it "client successfully authenticates with namespace label restriction" do
+      #API is removed
+      xit "client successfully authenticates with namespace label restriction" do
         define_and_grant_host(
           account: account,
           host_id: test_app_host,
@@ -349,7 +352,8 @@ describe AuthenticateController, :type => :request do
         expect(token.claims).to have_key('iat')
       end
 
-      it "client fails when given both namespace name and label restriction" do
+      #API is removed
+      xit "client fails when given both namespace name and label restriction" do
         define_and_grant_host(
           account: account,
           host_id: test_app_host,
@@ -377,7 +381,8 @@ describe AuthenticateController, :type => :request do
         }
       end
 
-      it "client fails when given a url that is not kubernetes server" do
+      #API is removed
+      xit "client fails when given a url that is not kubernetes server" do
         configure_k8s_api_access(
           account: account,
           service_id: service_id,